refactor(domain): migrate PathNodes coordinate properties from Float to Double
- Change all coordinate interfaces and PathNode subclass properties to Double.
- Update copy() parameter types and property initializers.
- Output format (f suffix) is preserved in buildParameters().

+++ svg-to-compose/src/commonMain/kotlin/dev/tonholo/s2c/domain/svg/SvgPolygonNode.kt
@@ -25,7 +25,7 @@ abstract class SvgGraphicNodeWithListOfPoints<T>(
                 .findAll(points)
                 .map { it.value }
                 .windowed(size = 2, step = 2)
-                .map { (x, y) -> Point2D(x.toFloat(), y.toFloat()) }
+                .map { (x, y) -> Point2D(x.toDouble(), y.toDouble()) }
                 .toList()
         }
     }
